refactor(Paper)
- rename to Card as in our design system
- remove un-necessary code for elevation
now our Header menu, Apps Menu, Menu itself looks standard
