+++ src/core/Card/styles.css
@@ -1,0 +1,9 @@
+.container {
+    position: relative;
+    height: 186px;
+    width: 358px;
+    border-radius: 2px;
+    background-color: #fff;
+    box-shadow: 0 0 2px 0 rgba(0, 0, 0, 0.14), 0 2px 2px 0 rgba(0, 0, 0, 0.12),
+        0 1px 3px 0 rgba(0, 0, 0, 0.2);
+}
