+++ src/core/Menu/Menu.js
@@ -3,21 +3,21 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import MenuItem from './MenuItem'
-import Paper from '../Paper'
+import Card from '../Card'
 
 import s from './styles'
 
 function Menu({ width, height, list, onSelect, onClose, children }) {
     if (children) {
         return (
-            <Paper elevation={4} width={width} height={height}>
+            <Card width={width} height={height}>
                 <ul className={s('container')}>{children}</ul>
-            </Paper>
+            </Card>
         )
     }
 
     return (
-        <Paper elevation={4} width={width} height={height}>
+        <Card width={width} height={height}>
             <ul className={s('container')}>
                 {list.map(({ onClick, ...rest }, idx) => (
                     <MenuItem
@@ -29,7 +29,7 @@ function Menu({ width, height, list, onSelect, onClose, children }) {
                     />
                 ))}
             </ul>
-        </Paper>
+        </Card>
     )
 }
