--- src/core/Paper/styles.css
@@ -1,41 +1,0 @@
-/** @format */
-
-.container {
-    position: relative;
-    margin: 0 auto 0;
-    border-radius: 2px;
-    background-color: #ffffff;
-}
-
-.elevation-0 {
-    box-shadow: none;
-}
-
-.elevation-1 {
-    box-shadow: 0 2px 1px -1px rgba(0, 0, 0, 0.2),
-        0 1px 1px 0 rgba(0, 0, 0, 0.14), 0 1px 3px 0 rgba(0, 0, 0, 0.12);
-}
-.elevation-2 {
-    box-shadow: 0 3px 1px -2px rgba(0, 0, 0, 0.2),
-        0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 1px 5px 0 rgba(0, 0, 0, 0.12);
-}
-.elevation-3 {
-    box-shadow: 0 3px 3px -2px rgba(0, 0, 0, 0.2),
-        0 3px 4px 0 rgba(0, 0, 0, 0.14), 0 1px 8px 0 rgba(0, 0, 0, 0.12);
-}
-.elevation-4 {
-    box-shadow: 0 2px 4px -1px rgba(0, 0, 0, 0.2),
-        0 4px 5px 0 rgba(0, 0, 0, 0.14), 0 1px 10px 0 rgba(0, 0, 0, 0.12);
-}
-.elevation-6 {
-    box-shadow: 0 3px 5px -1px rgba(0, 0, 0, 0.2),
-        0 6px 10px 0 rgba(0, 0, 0, 0.14), 0 1px 18px 0 rgba(0, 0, 0, 0.12);
-}
-.elevation-12 {
-    box-shadow: 0 7px 8px -4px rgba(0, 0, 0, 0.2),
-        0 12px 17px 2px rgba(0, 0, 0, 0.14), 0 5px 22px 4px rgba(0, 0, 0, 0.12);
-}
-.elevation-24 {
-    box-shadow: 0 11px 15px -7px rgba(0, 0, 0, 0.2),
-        0 24px 38px 3px rgba(0, 0, 0, 0.14), 0 9px 46px 8px rgba(0, 0, 0, 0.12);
-}
