+++ src/index.js
@@ -4,7 +4,7 @@ export * from './core/Button'
 export * from './core/Icon'
 export * from './core/Logo'
 export * from './core/Menu'
-export * from './core/Paper'
+export * from './core/Card'
 export * from './core/Popover'
 export * from './core/Tooltip'
 export * from './widgets/HeaderBar'
